refactor(ui): move crossChainUrlHelper into utils/ (EFI-957)
